# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Molecular Vision AR (MVAR) API Documentation",
        "description": "Dokumentasi RESTful API untuk Molecular Vision AR (MVAR) Backend. API ini digunakan oleh aplikasi mobile (Flutter / Unity) dan client lainnya.",
        "contact": {
            "name": "Tim Pengembang MVAR",
            "email": "dev@mvar.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Current API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/materials": {
            "get": {
                "tags": [
                    "Materi"
                ],
                "summary": "Daftar Materi Pembelajaran Kimia",
                "description": "Mengambil daftar ringkas materi pembelajaran kimia dengan opsi filter kategori dan pencarian judul materi.",
                "operationId": "fa952cf0966de3a090a9c0ad9b075b39",
                "parameters": [
                    {
                        "name": "id_category",
                        "in": "query",
                        "description": "Filter materi berdasarkan ID kategori kimia (misal: 1 untuk Stoikiometri)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Kata kunci pencarian judul atau deskripsi materi",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Hukum Dasar"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar materi berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Daftar materi berhasil diambil."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id_material": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "id_category": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Hukum Dasar Kimia & Konsep Mol"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Pengantar perhitungan kimia dan stoikiometri dasar."
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id_category": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Stoikiometri"
                                                            }
                                                        },
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau tidak disertakan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/materials/{id}": {
            "get": {
                "tags": [
                    "Materi"
                ],
                "summary": "Detail Konten Materi Pembelajaran",
                "description": "Mengambil detail lengkap materi pembelajaran kimia termasuk isi konten teori (HTML / teks).",
                "operationId": "4c9307cd61989b969d7eb250eaa5260f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID materi pembelajaran",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail materi berhasil diambil.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detail materi berhasil diambil."
                                        },
                                        "data": {
                                            "properties": {
                                                "id_material": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_category": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Hukum Dasar Kimia & Konsep Mol"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Pengantar perhitungan kimia dan stoikiometri dasar."
                                                },
                                                "content": {
                                                    "type": "string",
                                                    "example": "<h2>Konsep Mol</h2><p>Mol adalah satuan dasar dalam SI yang menyatakan jumlah zat...</p>"
                                                },
                                                "category": {
                                                    "properties": {
                                                        "id_category": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Stoikiometri"
                                                        }
                                                    },
                                                    "type": "object",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token tidak valid atau tidak disertakan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Materi pembelajaran tidak ditemukan.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Materi pembelajaran tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Gunakan token Sanctum dengan format: Bearer <token>",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}